requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
requests
psycopg2-binary